datable buttons and data deletion popups

+++ src/app/Pages/CatalogSecurity/view-catalogue-security-levels/view-catalogue-security-levels.component.ts
@@ -5,6 +5,7 @@ import { FormGroup, FormControl } from '@angular/forms';
 import { ToastrService } from 'ngx-toastr';
 import { HttpClient } from '@angular/common/http';
 import { Subject } from 'rxjs';
+import { NgPopupsService } from 'ng-popups';
 
 import { CatalogSecurity } from 'src/app/models/catalog-security.model';
 
@@ -20,7 +21,8 @@ export class ViewCatalogueSecurityLevelsComponent implements OnInit, OnDestroy {
   title!: string;
   role:any;
 
-  dtOptions: DataTables.Settings = {};
+  // dtOptions: DataTables.Settings = {};
+  dtOptions: any = {};
   dtTrigger: Subject<any> = new Subject<any>();
 
 
@@ -40,7 +42,8 @@ export class ViewCatalogueSecurityLevelsComponent implements OnInit, OnDestroy {
     private authservice: ApiPipeService,
     private router: Router,
     private toastr: ToastrService,
-    private http: HttpClient
+    private http: HttpClient,
+    private ngPopups: NgPopupsService,
   ) { }
 
   ngOnInit(): void {
@@ -56,6 +59,34 @@ export class ViewCatalogueSecurityLevelsComponent implements OnInit, OnDestroy {
     this.posts = [];
     this.users();
     this.initForm();
+    this.dtOptions = {
+      dom:'Bfrtip',
+      // dom:'Btp',
+      buttons: [
+        // 'columnsToggle',
+        // 'colvis',
+        {
+          extend:'copy',
+          tag: 'button',
+          className: "btn blue btn-outline"
+        },
+        {
+          extend:'print',
+          tag: 'button',
+          className: "btn yellow btn-outline"
+        },
+        {
+          extend:'excel',
+          tag: 'button',
+          className: "btn green btn-outline"
+        },
+        {
+          extend:'pdf',
+          tag: 'button',
+          className: "btn red btn-outline"
+        },
+      ]
+    }
   }
 
   ngOnDestroy(): void {
@@ -87,27 +118,41 @@ export class ViewCatalogueSecurityLevelsComponent implements OnInit, OnDestroy {
 
   onSelect(selectedItem: any) {
     this.id = selectedItem.CatalogSecurityFlag_id
-    console.log("Selected item Id: ", selectedItem.CatalogSecurityFlag_id);
-    this.http.delete('http://127.0.0.1:8899/apiv1/delete_catalog_security_flag/' + this.id)
-      .subscribe(response => {
-        this.deleteresp = response;
-        console.log(this.deleteresp.message)
-        if (this.deleteresp.message == "Catalog Security Flag successfully deleted.") {
-          this.toastr.success("Catalog Security Flag successfully deleted.", "", {
-            timeOut: 2000,
-            positionClass: 'toast-top-center',
-            progressBar: true,
-            progressAnimation: 'increasing'
-          })
-          setTimeout(() => {
-            this.authservice.reload();
-          }, 1000);
 
-        } else {
-          this.authservice.catalogSecurityStatusUpdate()
-        }
-        console.log(this.deleteresp)
-      });
+    this.ngPopups.confirm("Are you sure you want to delete ?",{
+      // theme: 'material',
+      color:'OrangeRed',
+      okButtonText: 'Yes',
+      cancelButtonText:'No',
+      title: "Confirm",
+    })
+    .subscribe(res => {
+      if (res) {
+        console.log("Selected item Id: ", selectedItem.CatalogSecurityFlag_id);
+        this.http.delete('http://127.0.0.1:8899/apiv1/delete_catalog_security_flag/' + this.id)
+          .subscribe(response => {
+            this.deleteresp = response;
+            console.log(this.deleteresp.message)
+            if (this.deleteresp.message == "Catalog Security Flag successfully deleted.") {
+              this.toastr.success("Catalog Security Flag successfully deleted.", "", {
+                timeOut: 2000,
+                positionClass: 'toast-top-center',
+                progressBar: true,
+                progressAnimation: 'increasing'
+              })
+              setTimeout(() => {
+                this.authservice.reload();
+              }, 1000);
+
+            } else {
+              this.authservice.catalogSecurityStatusUpdate()
+            }
+            console.log(this.deleteresp)
+          });
+      } else {
+        console.log("You clicked cancel.")
+      }
+    });
   }
 
   onSelectEdit(selectedItem: any) {
