+++ src/app/Pages/CoreTypes/view-core-type/view-core-type.component.ts
@@ -5,6 +5,8 @@ import { FormGroup, FormControl } from '@angular/forms';
 import { ToastrService } from 'ngx-toastr';
 import { HttpClient } from '@angular/common/http';
 import { Subject } from 'rxjs';
+import { NgPopupsService } from 'ng-popups';
+
 
 import { CoreType } from 'src/app/models/core-type.model';
 
@@ -13,13 +15,15 @@ import { CoreType } from 'src/app/models/core-type.model';
   templateUrl: './view-core-type.component.html',
   styleUrls: ['./view-core-type.component.css']
 })
+
 export class ViewCoreTypeComponent implements OnInit, OnDestroy {
 
   formGroup!: FormGroup;
   title!: string;
   role:any;
 
-  dtOptions: DataTables.Settings = {};
+  // dtOptions: DataTables.Settings = {};
+  dtOptions: any = {};
   dtTrigger: Subject<any> = new Subject<any>();
 
 
@@ -39,7 +43,9 @@ export class ViewCoreTypeComponent implements OnInit, OnDestroy {
     private authservice: ApiPipeService,
     private router: Router,
     private toastr: ToastrService,
-    private http: HttpClient
+    private http: HttpClient,
+    private ngPopups: NgPopupsService,
+
   ) { }
 
   ngOnInit(): void {
@@ -55,6 +61,34 @@ export class ViewCoreTypeComponent implements OnInit, OnDestroy {
     this.posts = [];
     this.users();
     this.initForm();
+    this.dtOptions = {
+      dom:'Bfrtip',
+      // dom:'Btp',
+      buttons: [
+        // 'columnsToggle',
+        // 'colvis',
+        {
+          extend:'copy',
+          tag: 'button',
+          className: "btn blue btn-outline"
+        },
+        {
+          extend:'print',
+          tag: 'button',
+          className: "btn yellow btn-outline"
+        },
+        {
+          extend:'excel',
+          tag: 'button',
+          className: "btn green btn-outline"
+        },
+        {
+          extend:'pdf',
+          tag: 'button',
+          className: "btn red btn-outline"
+        },
+      ]
+    }
   }
 
   ngOnDestroy(): void {
@@ -86,27 +120,41 @@ export class ViewCoreTypeComponent implements OnInit, OnDestroy {
 
   onSelect(selectedItem: any) {
     this.id = selectedItem.CoreType_id
-    console.log("Selected item Id: ", selectedItem.CoreType_id);
-    this.http.delete('http://127.0.0.1:8899/apiv1/delete_core_type/' + this.id)
-      .subscribe(response => {
-        this.deleteresp = response;
-        console.log(this.deleteresp.message)
-        if (this.deleteresp.message == "Core Type successfully deleted.") {
-          this.toastr.success("Core Type successfully deleted.", "", {
-            timeOut: 2000,
-            positionClass: 'toast-top-center',
-            progressBar: true,
-            progressAnimation: 'increasing'
-          })
-          setTimeout(() => {
-            this.authservice.reload();
-          }, 1000);
 
-        } else {
-          this.authservice.coreTypeStatusUpdate()
-        }
-        console.log(this.deleteresp)
-      });
+    this.ngPopups.confirm("Are you sure you want to delete ?",{
+      // theme: 'material',
+      color:'OrangeRed',
+      okButtonText: 'Yes',
+      cancelButtonText:'No',
+      title: "Confirm",
+    })
+    .subscribe(res => {
+      if (res) {
+        console.log("Selected item Id: ", selectedItem.CoreType_id);
+        this.http.delete('http://127.0.0.1:8899/apiv1/delete_core_type/' + this.id)
+          .subscribe(response => {
+            this.deleteresp = response;
+            console.log(this.deleteresp.message)
+            if (this.deleteresp.message == "Core Type successfully deleted.") {
+              this.toastr.success("Core Type successfully deleted.", "", {
+                timeOut: 2000,
+                positionClass: 'toast-top-center',
+                progressBar: true,
+                progressAnimation: 'increasing'
+              })
+              setTimeout(() => {
+                this.authservice.reload();
+              }, 1000);
+
+            } else {
+              this.authservice.coreTypeStatusUpdate()
+            }
+            console.log(this.deleteresp)
+          });
+      } else {
+        console.log("You clicked cancel.")
+      }
+    });
   }
 
   onSelectEdit(selectedItem: any) {
