+++ src/app/Pages/Web-security/web-security-levels/web-security-levels.component.ts
@@ -5,6 +5,8 @@ import { FormGroup, FormControl } from '@angular/forms';
 import { ToastrService } from 'ngx-toastr';
 import { HttpClient } from '@angular/common/http';
 import { Subject } from 'rxjs';
+import { NgPopupsService } from 'ng-popups';
+
 
 
 import { WebSecurity } from 'src/app/models/web-security.model';
@@ -14,12 +16,14 @@ import { WebSecurity } from 'src/app/models/web-security.model';
   templateUrl: './web-security-levels.component.html',
   styleUrls: ['./web-security-levels.component.css']
 })
+
 export class WebSecurityLevelsComponent implements OnInit, OnDestroy {
   formGroup!: FormGroup;
   title!: string;
   role:any;
 
-  dtOptions: DataTables.Settings = {};
+  // dtOptions: DataTables.Settings = {};
+  dtOptions: any = {};
   dtTrigger: Subject<any> = new Subject<any>();
 
 
@@ -39,7 +43,9 @@ export class WebSecurityLevelsComponent implements OnInit, OnDestroy {
     private authservice: ApiPipeService,
     private router: Router,
     private toastr: ToastrService,
-    private http: HttpClient
+    private http: HttpClient,
+    private ngPopups: NgPopupsService,
+
   ) { }
 
   ngOnInit(): void {
@@ -51,10 +57,38 @@ export class WebSecurityLevelsComponent implements OnInit, OnDestroy {
     }else{
     this.role= false;
     }
-  
+
     this.posts = [];
     this.users();
     this.initForm();
+    this.dtOptions = {
+      dom:'Bfrtip',
+      // dom:'Btp',
+      buttons: [
+        // 'columnsToggle',
+        // 'colvis',
+        {
+          extend:'copy',
+          tag: 'button',
+          className: "btn blue btn-outline"
+        },
+        {
+          extend:'print',
+          tag: 'button',
+          className: "btn yellow btn-outline"
+        },
+        {
+          extend:'excel',
+          tag: 'button',
+          className: "btn green btn-outline"
+        },
+        {
+          extend:'pdf',
+          tag: 'button',
+          className: "btn red btn-outline"
+        },
+      ]
+    }
   }
 
   ngOnDestroy(): void {
@@ -86,27 +120,42 @@ export class WebSecurityLevelsComponent implements OnInit, OnDestroy {
 
   onSelect(selectedItem: any) {
     this.id = selectedItem.WebSecurityLevel_id
-    console.log("Selected item Id: ", selectedItem.WebSecurityLevel_id);
-    this.http.delete('http://127.0.0.1:8899/apiv1/delete_web_security_level/' + this.id)
-      .subscribe(response => {
-        this.deleteresp = response;
-        console.log(this.deleteresp.message)
-        if (this.deleteresp.message == "Web security level successfully deleted.") {
-          this.toastr.success("Web security level successfully deleted.", "", {
-            timeOut: 2000,
-            positionClass: 'toast-top-center',
-            progressBar: true,
-            progressAnimation: 'increasing'
-          })
-          setTimeout(() => {                           
-            this.authservice.reload();
-          }, 1000);
 
-        } else {
-          this.authservice.securityStatusUpdate()
-        }
-        console.log(this.deleteresp)
-      });
+    this.ngPopups.confirm("Are you sure you want to delete ?",{
+      // theme: 'material',
+      color:'OrangeRed',
+      okButtonText: 'Yes',
+      cancelButtonText:'No',
+      title: "Confirm",
+    })
+    .subscribe(res => {
+      if (res) {
+      console.log("Selected item Id: ", selectedItem.WebSecurityLevel_id);
+      this.http.delete('http://127.0.0.1:8899/apiv1/delete_web_security_level/' + this.id)
+        .subscribe(response => {
+          this.deleteresp = response;
+          console.log(this.deleteresp.message)
+          if (this.deleteresp.message == "Web security level successfully deleted.") {
+            this.toastr.success("Web security level successfully deleted.", "", {
+              timeOut: 2000,
+              positionClass: 'toast-top-center',
+              progressBar: true,
+              progressAnimation: 'increasing'
+            })
+            setTimeout(() => {
+              this.authservice.reload();
+            }, 1000);
+
+          } else {
+            this.authservice.securityStatusUpdate()
+          }
+          console.log(this.deleteresp)
+        });
+      } else {
+        console.log("You clicked cancel.")
+      }
+    });
+
   }
 
   onSelectEdit(selectedItem: any) {
@@ -153,10 +202,10 @@ export class WebSecurityLevelsComponent implements OnInit, OnDestroy {
             progressAnimation: 'increasing'
           })
           this.formGroup.reset();
-          setTimeout(() => {                           
+          setTimeout(() => {
             this.authservice.reload();
           }, 1000);
-          
+
 
         } else {
           this.authservice.securityStatusUpdate()
@@ -190,6 +239,3 @@ export class WebSecurityLevelsComponent implements OnInit, OnDestroy {
 
 
 }
-
-
-
